@@ -3,15 +3,15 @@
 # BSD-style license that can be found in the LICENSE file.
 
 name: ffigen
-version: 4.1.2
+version: 5.0.0-dev.0
 homepage: https://github.com/dart-lang/ffigen
 description: Generator for FFI bindings, using LibClang to parse C header files.
 
 environment:
-  sdk: '>=2.14.0 <3.0.0'
+  sdk: '>=2.16.0-124.0.dev <3.0.0'
 
 dependencies:
-  ffi: ^1.0.0
+  ffi: ^1.2.0-dev.0
   yaml: ^3.0.0
   path: ^1.8.0
   quiver: ^3.0.0
