@@ -2,6 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'package:ffigen/src/code_generator/imports.dart';
 import 'package:ffigen/src/code_generator/utils.dart';
 
 import 'binding.dart';
@@ -22,8 +23,15 @@ class Writer {
   late String _className;
   final String? classDocComment;
 
-  late String _ffiLibraryPrefix;
-  String get ffiLibraryPrefix => _ffiLibraryPrefix;
+  String? _ffiLibraryPrefix;
+  String get ffiLibraryPrefix {
+    _ffiLibraryPrefix ??= libraryImports
+        .firstWhere((element) => element.name == ffiImport.name)
+        .prefix;
+    return _ffiLibraryPrefix!;
+  }
+
+  final Set<LibraryImport> libraryImports;
 
   late String _lookupFuncIdentifier;
   String get lookupFuncIdentifier => _lookupFuncIdentifier;
@@ -55,6 +63,7 @@ class Writer {
     required this.noLookUpBindings,
     required String className,
     required this.dartBool,
+    required this.libraryImports,
     this.classDocComment,
     this.header,
   }) {
@@ -75,12 +84,17 @@ class Writer {
       markUsed: [_initialWrapperLevelUniqueNamer, _initialTopLevelUniqueNamer],
     );
 
-    /// [_ffiLibraryPrefix] should be unique unique among all names.
-    _ffiLibraryPrefix = _resolveNameConflict(
-      name: 'ffi',
-      makeUnique: allLevelsUniqueNamer,
-      markUsed: [_initialWrapperLevelUniqueNamer, _initialTopLevelUniqueNamer],
-    );
+    /// Library imports prefix should be unique unique among all names.
+    for (final lib in libraryImports) {
+      lib.prefix = _resolveNameConflict(
+        name: lib.prefix,
+        makeUnique: allLevelsUniqueNamer,
+        markUsed: [
+          _initialWrapperLevelUniqueNamer,
+          _initialTopLevelUniqueNamer
+        ],
+      );
+    }
 
     /// [_lookupFuncIdentifier] should be unique in top level.
     _lookupFuncIdentifier = _resolveNameConflict(
@@ -161,8 +175,10 @@ class Writer {
         'AUTO GENERATED FILE, DO NOT EDIT.\n\nGenerated by `package:ffigen`.'));
 
     // Write neccesary imports.
-    s.write("import 'dart:ffi' as $ffiLibraryPrefix;\n");
-    s.write('\n');
+    for (final lib in libraryImports) {
+      s.write("import '${lib.importPath}' as ${lib.prefix};");
+      s.write('\n');
+    }
 
     /// Write [lookUpBindings].
     if (lookUpBindings.isNotEmpty) {
