@@ -2,6 +2,7 @@
 //
 // Generated by `package:ffigen`.
 import 'dart:ffi' as ffi;
+import 'package:ffi/ffi.dart' as pkg_ffi;
 
 /// Bindings to `headers/example.h`.
 class NativeLibrary {
@@ -30,14 +31,14 @@ class NativeLibrary {
     );
   }
 
-  late final _sumPtr =
-      _lookup<ffi.NativeFunction<ffi.Int32 Function(ffi.Int32, ffi.Int32)>>(
-          'sum');
+  late final _sumPtr = _lookup<
+          ffi.NativeFunction<pkg_ffi.Int Function(pkg_ffi.Int, pkg_ffi.Int)>>(
+      'sum');
   late final _sum = _sumPtr.asFunction<int Function(int, int)>();
 
   /// Subtracts 2 integers.
   int subtract(
-    ffi.Pointer<ffi.Int32> a,
+    ffi.Pointer<pkg_ffi.Int> a,
     int b,
   ) {
     return _subtract(
@@ -48,12 +49,13 @@ class NativeLibrary {
 
   late final _subtractPtr = _lookup<
       ffi.NativeFunction<
-          ffi.Int32 Function(ffi.Pointer<ffi.Int32>, ffi.Int32)>>('subtract');
+          pkg_ffi.Int Function(
+              ffi.Pointer<pkg_ffi.Int>, pkg_ffi.Int)>>('subtract');
   late final _subtract =
-      _subtractPtr.asFunction<int Function(ffi.Pointer<ffi.Int32>, int)>();
+      _subtractPtr.asFunction<int Function(ffi.Pointer<pkg_ffi.Int>, int)>();
 
   /// Multiplies 2 integers, returns pointer to an integer,.
-  ffi.Pointer<ffi.Int32> multiply(
+  ffi.Pointer<pkg_ffi.Int> multiply(
     int a,
     int b,
   ) {
@@ -65,9 +67,10 @@ class NativeLibrary {
 
   late final _multiplyPtr = _lookup<
       ffi.NativeFunction<
-          ffi.Pointer<ffi.Int32> Function(ffi.Int32, ffi.Int32)>>('multiply');
+          ffi.Pointer<pkg_ffi.Int> Function(
+              pkg_ffi.Int, pkg_ffi.Int)>>('multiply');
   late final _multiply =
-      _multiplyPtr.asFunction<ffi.Pointer<ffi.Int32> Function(int, int)>();
+      _multiplyPtr.asFunction<ffi.Pointer<pkg_ffi.Int> Function(int, int)>();
 
   /// Divides 2 integers, returns pointer to a float.
   ffi.Pointer<ffi.Float> divide(
@@ -82,7 +85,7 @@ class NativeLibrary {
 
   late final _dividePtr = _lookup<
       ffi.NativeFunction<
-          ffi.Pointer<ffi.Float> Function(ffi.Int32, ffi.Int32)>>('divide');
+          ffi.Pointer<ffi.Float> Function(pkg_ffi.Int, pkg_ffi.Int)>>('divide');
   late final _divide =
       _dividePtr.asFunction<ffi.Pointer<ffi.Float> Function(int, int)>();
 
