@@ -2,24 +2,25 @@
 //
 // Generated by `package:ffigen`.
 import 'dart:ffi' as ffi;
+import 'package:ffi/ffi.dart' as pkg_ffi;
 
 class Struct1 extends ffi.Struct {
-  @ffi.Int8()
+  @pkg_ffi.Char()
   external int a;
 }
 
 class Union1 extends ffi.Union {
-  @ffi.Int8()
+  @pkg_ffi.Char()
   external int a;
 }
 
 class EmptyUnion extends ffi.Opaque {}
 
 class Primitives extends ffi.Union {
-  @ffi.Int8()
+  @pkg_ffi.Char()
   external int a;
 
-  @ffi.Int32()
+  @pkg_ffi.Int()
   external int b;
 
   @ffi.Float()
@@ -30,7 +31,7 @@ class Primitives extends ffi.Union {
 }
 
 class PrimitivesWithPointers extends ffi.Union {
-  @ffi.Int8()
+  @pkg_ffi.Char()
   external int a;
 
   @ffi.Float()
@@ -45,7 +46,7 @@ class PrimitivesWithPointers extends ffi.Union {
 
 class WithArray extends ffi.Union {
   @ffi.Array.multi([10])
-  external ffi.Array<ffi.Int8> a;
+  external ffi.Array<pkg_ffi.Char> a;
 
   @ffi.Array.multi([10])
   external ffi.Array<Union1> b;
