@@ -4,17 +4,18 @@
 //
 // Generated by `package:ffigen`.
 import 'dart:ffi' as ffi;
+import 'package:ffi/ffi.dart' as pkg_ffi;
 
 class A extends ffi.Struct {
-  @ffi.Int32()
+  @pkg_ffi.Int()
   external int a;
 }
 
 class B extends ffi.Struct {
-  @ffi.Int32()
+  @pkg_ffi.Int()
   external int B1;
 
-  @ffi.Int32()
+  @pkg_ffi.Int()
   external int A;
 }
 
