@@ -27,8 +27,12 @@ ${strings.headers}:
   ${strings.includeDirectives}:
     - '**cJSON.h'
 ${strings.comments}: false
-${strings.typedefmap}:
-  'size_t': 'IntPtr'
+${strings.typeMap}:
+  ${strings.typeMapTypedefs}:
+    'size_t':
+      lib: 'pkg_ffi'
+      c-type: 'Size'
+      dart-type: 'int'
 ${strings.preamble}: |
   // Copyright (c) 2009-2017 Dave Gamble and cJSON contributors
   //
