@@ -2,6 +2,7 @@
 //
 // Generated by `package:ffigen`.
 import 'dart:ffi' as ffi;
+import 'package:ffi/ffi.dart' as pkg_ffi;
 
 /// Comment Markup Test
 class NativeLibrary {
@@ -54,7 +55,7 @@ class NativeLibrary {
 class Com4 extends ffi.Struct {
   /// Muli-line test comment for struct field
   /// With multiple line and both // and ///.
-  @ffi.Int32()
+  @pkg_ffi.Int()
   external int a;
 
   /// Single line field comment.
