@@ -2,6 +2,7 @@
 //
 // Generated by `package:ffigen`.
 import 'dart:ffi' as ffi;
+import 'package:ffi/ffi.dart' as pkg_ffi;
 
 /// Forward Declaration Test
 class NativeLibrary {
@@ -36,10 +37,10 @@ class NativeLibrary {
 }
 
 class A extends ffi.Struct {
-  @ffi.Int32()
+  @pkg_ffi.Int()
   external int a;
 
-  @ffi.Int32()
+  @pkg_ffi.Int()
   external int b;
 }
 
