@@ -2,6 +2,7 @@
 //
 // Generated by `package:ffigen`.
 import 'dart:ffi' as ffi;
+import 'package:ffi/ffi.dart' as pkg_ffi;
 
 /// Opaque Dependencies Test
 class NativeLibrary {
@@ -63,7 +64,7 @@ class E extends ffi.Struct {
 class C extends ffi.Opaque {}
 
 class D extends ffi.Struct {
-  @ffi.Int32()
+  @pkg_ffi.Int()
   external int a;
 }
 
@@ -80,6 +81,6 @@ class UE extends ffi.Union {
 class UC extends ffi.Opaque {}
 
 class UD extends ffi.Union {
-  @ffi.Int32()
+  @pkg_ffi.Int()
   external int a;
 }
