@@ -2,48 +2,49 @@
 //
 // Generated by `package:ffigen`.
 import 'dart:ffi' as ffi;
+import 'package:ffi/ffi.dart' as pkg_ffi;
 
 class NormalStruct1 extends ffi.Struct {
-  @ffi.Int8()
+  @pkg_ffi.Char()
   external int a;
 }
 
 /// Should not be packed.
 class StructWithAttr extends ffi.Struct {
-  external ffi.Pointer<ffi.Int32> a;
+  external ffi.Pointer<pkg_ffi.Int> a;
 
-  external ffi.Pointer<ffi.Int32> b;
+  external ffi.Pointer<pkg_ffi.Int> b;
 }
 
 /// Should be packed with 1.
 @ffi.Packed(1)
 class PackedAttr extends ffi.Struct {
-  @ffi.Int32()
+  @pkg_ffi.Int()
   external int a;
 }
 
 /// Should be packed with 8.
 @ffi.Packed(8)
 class PackedAttrAlign8 extends ffi.Struct {
-  @ffi.Int32()
+  @pkg_ffi.Int()
   external int a;
 }
 
 /// Should be packed with 2.
 @ffi.Packed(2)
 class Pack2WithPragma extends ffi.Struct {
-  @ffi.Int32()
+  @pkg_ffi.Int()
   external int a;
 }
 
 /// Should be packed with 4.
 @ffi.Packed(4)
 class Pack4WithPragma extends ffi.Struct {
-  @ffi.Int64()
+  @pkg_ffi.LongLong()
   external int a;
 }
 
 class NormalStruct2 extends ffi.Struct {
-  @ffi.Int8()
+  @pkg_ffi.Char()
   external int a;
 }
