@@ -4,6 +4,7 @@
 //
 // Generated by `package:ffigen`.
 import 'dart:ffi' as ffi;
+import 'package:ffi/ffi.dart' as pkg_ffi;
 
 class Bindings {
   /// Holds the symbol lookup function.
@@ -21,7 +22,7 @@ class Bindings {
 
   NamedFunctionProto func1(
     NamedFunctionProto named,
-    ffi.Pointer<ffi.NativeFunction<ffi.Void Function(ffi.Int32)>> unnamed,
+    ffi.Pointer<ffi.NativeFunction<ffi.Void Function(pkg_ffi.Int)>> unnamed,
   ) {
     return _func1(
       named,
@@ -30,14 +31,15 @@ class Bindings {
   }
 
   late final _func1Ptr = _lookup<
-      ffi.NativeFunction<
-          NamedFunctionProto Function(
-              NamedFunctionProto,
-              ffi.Pointer<
-                  ffi.NativeFunction<ffi.Void Function(ffi.Int32)>>)>>('func1');
+          ffi.NativeFunction<
+              NamedFunctionProto Function(
+                  NamedFunctionProto,
+                  ffi.Pointer<
+                      ffi.NativeFunction<ffi.Void Function(pkg_ffi.Int)>>)>>(
+      'func1');
   late final _func1 = _func1Ptr.asFunction<
       NamedFunctionProto Function(NamedFunctionProto,
-          ffi.Pointer<ffi.NativeFunction<ffi.Void Function(ffi.Int32)>>)>();
+          ffi.Pointer<ffi.NativeFunction<ffi.Void Function(pkg_ffi.Int)>>)>();
 
   void func2(
     ffi.Pointer<NTyperef1> arg0,
