@@ -2,6 +2,7 @@
 //
 // Generated by `package:ffigen`.
 import 'dart:ffi' as ffi;
+import 'package:ffi/ffi.dart' as pkg_ffi;
 
 /// Unions Test
 class NativeLibrary {
@@ -49,7 +50,7 @@ class NativeLibrary {
 }
 
 class Union1 extends ffi.Union {
-  @ffi.Int32()
+  @pkg_ffi.Int()
   external int a;
 }
 
