@@ -4,6 +4,7 @@
 //
 // Generated by `package:ffigen`.
 import 'dart:ffi' as ffi;
+import 'package:ffi/ffi.dart' as pkg_ffi;
 
 /// Native tests.
 class NativeLibrary {
@@ -211,8 +212,8 @@ class NativeLibrary {
 
   late final _Function1StructReturnByValuePtr = _lookup<
       ffi.NativeFunction<
-          Struct3 Function(ffi.Int32, ffi.Int32,
-              ffi.Int32)>>('Function1StructReturnByValue');
+          Struct3 Function(pkg_ffi.Int, pkg_ffi.Int,
+              pkg_ffi.Int)>>('Function1StructReturnByValue');
   late final _Function1StructReturnByValue = _Function1StructReturnByValuePtr
       .asFunction<Struct3 Function(int, int, int)>();
 
@@ -225,7 +226,7 @@ class NativeLibrary {
   }
 
   late final _Function1StructPassByValuePtr =
-      _lookup<ffi.NativeFunction<ffi.Int32 Function(Struct3)>>(
+      _lookup<ffi.NativeFunction<pkg_ffi.Int Function(Struct3)>>(
           'Function1StructPassByValue');
   late final _Function1StructPassByValue =
       _Function1StructPassByValuePtr.asFunction<int Function(Struct3)>();
@@ -240,12 +241,12 @@ class Struct1 extends ffi.Struct {
 }
 
 class Struct3 extends ffi.Struct {
-  @ffi.Int32()
+  @pkg_ffi.Int()
   external int a;
 
-  @ffi.Int32()
+  @pkg_ffi.Int()
   external int b;
 
-  @ffi.Int32()
+  @pkg_ffi.Int()
   external int c;
 }
